ASoC: msm: q6: update render flag to NON_DMA for handling underrun
In scenarios where the driver is in underrun state, restart
mechanism triggers when the render flag is in NON_DMA mode. The
compressed driver which is opened in mmap mode runs into underrun
scenarios when the valid buffer length is smaller than the
buffering in the DSP. Update render flag accordingly so that the
restart is trigger when driver is in underrun state.

CRs-Fixed: 487732
Change-Id: Ia4ee5dcf1d079ebef69e9ee627591bc039934add
Signed-off-by: Jitendra Singh Naruka <jnaruk@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>