usb: usb_bam: set activity state upon hsic consumer request
In the original design, upon hsic consumer request, the hsic core was
resumed, but the hsic bam state stayed in inactivity, waiting for a
packet to arrive through the bam pipes in order to get back to activity
state, ask for the hsic producer and notify the USB class driver.

However, since the resume of the hsic core upon hsic consumer request,
the USB class driver is notified about the resume by the ehci resume
callback. Therefore if we will wait for activity state the USB class
driver will get also the activity indication after the resume. On both
indications the class driver set a pm runtime vote. Therefore this
scenario lead to a double pm runtime vote that will prevent hsic suspend
from that point on.

Fix this by enter to activity state right when getting hsic consumer
request, therefore requesting the hsic producer and wait for inactivity
timer interrupt from the bam instead of activity wakeup interrupt.

Signed-off-by: Ido Shayevitz <idos@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>